dedup getRoomData in DataGetter.kt

+++ app/src/main/java/ch/sdp/vibester/database/DataGetter.kt
@@ -298,8 +298,8 @@ class DataGetter @Inject constructor() {
                         }
                         val gameSize: Int = ((snapshot.value as Map<String, Object>)["gameSize"] as Long).toInt()
                         val gameMode: String = (snapshot.value as Map<String, Object>)["gameMode"] as String
-                        val difficultyLevel: Int = ((snapshot.value as Map<String, Object>)["difficulty"] as Long).toInt()
-                        songListCallback(gameSongList, gameSize, gameMode, difficultyLevel)
+                        val difficultyLevel: Long = (snapshot.value as Map<String, Object>)["difficulty"] as Long
+                        songListCallback(gameSongList, gameSize, gameMode, difficultyLevel.toInt())
                     }
                 }
